refactored backend, removed old server code
