+++ src/server/src/routes/index.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import authRoutes from "./auth";
+
+const router = Router();
+
+router.use("/auth", authRoutes);
+
+export default router;
