+++ src/server/src/services/auth/jwtStrategy.ts
@@ -1,6 +1,6 @@
 import passport from "passport";
 import { ExtractJwt, Strategy as JwtStrategy } from "passport-jwt";
-import User from "../models/user.model";
+import User from "../../models/user";
 
 const isProduction = process.env.NODE_ENV === "production";
 const secretOrKey = isProduction ? process.env.JWT_SECRET_PROD : process.env.JWT_SECRET_DEV;
