--- src/server2/README.md
@@ -1,3 +1,0 @@
-# Node.js + Express.js
-
-W.I.P
