--- src/server2/babel.config.json
@@ -1,7 +1,0 @@
-{
-  "env": {
-    "test": {
-      "plugins": ["@babel/plugin-transform-modules-commonjs"]
-    }
-  }
-}
