--- src/server2/jest.config.js
@@ -1,5 +1,0 @@
-export default {
-  transform: {
-    "^.+\\.[t|j]sx?$": "babel-jest",
-  },
-};
