--- src/server2/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "W.I.P",
-  "main": "./src/server.js",
-  "type": "module",
-  "scripts": {
-    "start": "nodemon",
-    "dev": "nodemon",
-    "test": "jest --detectOpenHandles"
-  },
-  "keywords": [],
-  "author": "Harsh Keswani",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/core": "^7.23.2",
-    "@babel/plugin-transform-modules-commonjs": "^7.23.0",
-    "axios": "^1.5.1",
-    "babel-jest": "^29.7.0",
-    "bcrypt": "^5.1.1",
-    "body-parser": "^1.20.2",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "jest": "^29.7.0",
-    "jsonwebtoken": "^9.0.2",
-    "nodemon": "^3.0.1",
-    "pg": "^8.11.3",
-    "supertest": "^6.3.3"
-  },
-  "devDependencies": {
-    "@types/cookie-parser": "^1.4.4",
-    "eslint": "^8.51.0"
-  }
-}
